Don't install COPYING at top level of wheel
Telling Poetry to install the `COPYING` file without qualification
causes it to install that file at the very top level of the wheel,
outside the `shortuuid/` directory, and thus to potentially clash with
other packages.

We only need to explicitly install it in the sdist.  The wheel still
ends up with a copy of the file in `shortuuid-*.dist-info/COPYING`, so
no information is lost.